Update Bootstrap and Sass versions and refactor styles
Updated Bootstrap from version 5.1.3 to 5.3.2 and Sass from version 1.49.7 to 1.68.0 to utilize new features and improvements. Moreover, refactored the styles. bootstrap.scss was removed and a new _bootstrapComponents.scss was added to separate Bootstrap components and make them more manageable. The boilerplate styling in styles.scss was replaced with imports of the needed Bootstrap components. This makes it more clear which styles come from Bootstrap and which are custom.